refactor: split json upgrader into files
