+++ src/utils/type-transform/BackupFormat/JSON/version-0.ts
@@ -1,3 +1,4 @@
+/* eslint-disable import/no-deprecated */
 /**
  * @deprecated History JSON backup file
  */
