+++ src/utils/type-transform/BackupFormat/JSON/version-1.ts
@@ -1,5 +1,9 @@
+/* eslint-disable import/no-deprecated */
+import { BackupJSONFileVersion0 } from './version-0'
+import { ProfileIdentifier } from '../../../../database/type'
+
 /**
- * Latest JSON backup format
+ * @deprecated The old version 1 backup file before persona db was done.
  */
 export interface BackupJSONFileVersion1 {
     maskbookVersion: string
@@ -29,4 +33,35 @@ export function isBackupJSONFileVersion1(obj: object): obj is BackupJSONFileVers
     if (!Array.isArray(data.whoami)) return false
     if (!data.whoami) return false
     return true
+}
+
+// Since 8/21/2019, every backup file of version 1 should have grantedHostPermissions
+// Before 8/21/2019, we only support facebook, so we can auto upgrade the backup file
+const facebookHost = ['https://m.facebook.com/*', 'https://www.facebook.com/*']
+export function patchNonBreakingUpgradeForBackupJSONFileVersion1(json: BackupJSONFileVersion1): BackupJSONFileVersion1 {
+    if (json.grantedHostPermissions === undefined) {
+        json.grantedHostPermissions = facebookHost
+        json.maskbookVersion = '1.5.2'
+    }
+    if (!json.maskbookVersion) json.maskbookVersion = '1.6.0'
+    return json
+}
+export function upgradeFromBackupJSONFileVersion0(
+    json: BackupJSONFileVersion0,
+    identity: ProfileIdentifier,
+): BackupJSONFileVersion1 | null {
+    return {
+        maskbookVersion: '1.3.2',
+        version: 1,
+        whoami: [
+            {
+                localKey: json.local,
+                network: 'facebook.com',
+                publicKey: json.key.key.publicKey,
+                privateKey: json.key.key.privateKey!,
+                userId: identity.userId || '$self',
+            },
+        ],
+        grantedHostPermissions: facebookHost,
+    }
 }
