+++ src/utils/type-transform/BackupFormat/JSON/version-2.ts
@@ -1,1 +1,8 @@
+/* eslint-disable import/no-deprecated */
+/**
+ * @see https://github.com/DimensionDev/Maskbook/issues/194
+ *
+ * @deprecated Since this is the latest format, you should not use it.
+ * Use BackupJSONFileVersionLatest please.
+ */
 export interface BackupJSONFileVersion2 {}
